@@ -3,9 +3,9 @@ package edu.tamu.cap.config;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.messaging.simp.config.ChannelRegistration;
 import org.springframework.messaging.simp.config.MessageBrokerRegistry;
-import org.springframework.web.socket.config.annotation.AbstractWebSocketMessageBrokerConfigurer;
 import org.springframework.web.socket.config.annotation.EnableWebSocketMessageBroker;
 import org.springframework.web.socket.config.annotation.StompEndpointRegistry;
+import org.springframework.web.socket.config.annotation.WebSocketMessageBrokerConfigurer;
 import org.springframework.web.socket.config.annotation.WebSocketTransportRegistration;
 
 /**
@@ -16,7 +16,7 @@ import org.springframework.web.socket.config.annotation.WebSocketTransportRegist
  */
 @Configuration
 @EnableWebSocketMessageBroker
-public class AppWebSocketConfig extends AbstractWebSocketMessageBrokerConfigurer {
+public class AppWebSocketConfig implements WebSocketMessageBrokerConfigurer {
 
     /**
      * {@inheritDoc}
@@ -51,7 +51,7 @@ public class AppWebSocketConfig extends AbstractWebSocketMessageBrokerConfigurer
      */
     @Override
     public void registerStompEndpoints(StompEndpointRegistry registry) {
-        registry.addEndpoint("/connect").setAllowedOrigins("*").withSockJS();
+        registry.addEndpoint("/connect").setAllowedOriginPatterns("*").withSockJS();
     }
 
 }
