@@ -6,14 +6,15 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
 
-import org.apache.log4j.Logger;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 
 @Service
 public class TransactionService {
 
-    private static final Logger logger = Logger.getLogger(TransactionService.class);
+    private Logger logger = LoggerFactory.getLogger(this.getClass());
 
     private static final Map<String, ZonedDateTime> transactions = new ConcurrentHashMap<String, ZonedDateTime>();
 
