@@ -3,7 +3,7 @@ package edu.tamu.cap.ui.controller;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import org.springframework.boot.autoconfigure.web.ErrorController;
+import org.springframework.boot.web.servlet.error.ErrorController;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.ModelAndView;
@@ -13,11 +13,6 @@ import org.springframework.web.servlet.ModelAndView;
 public class Html5ModeErrorController implements ErrorController {
 
     private static final String PATH = "/error";
-
-    @Override
-    public String getErrorPath() {
-        return PATH;
-    }
 
     @RequestMapping(value = PATH)
     public ModelAndView error(HttpServletRequest request, HttpServletResponse response) {
