@@ -1,6 +1,6 @@
 package edu.tamu.cap.controller.repositoryviewcontext;
 
-import static org.mockito.Matchers.any;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.when;
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
@@ -16,6 +16,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 import org.aspectj.lang.ProceedingJoinPoint;
 import org.junit.jupiter.api.AfterEach;
@@ -93,8 +94,8 @@ public class RepositoryViewContextControllerTest {
         mockRepositoryView.setUsername("");
         mockRepositoryView.setPassword("");
 
-        when(repositoryViewRepo.getOne(1L)).thenReturn(mockRepositoryView);
-        when(repositoryViewRepo.findOne(1L)).thenReturn(mockRepositoryView);
+        when(repositoryViewRepo.getById(1L)).thenReturn(mockRepositoryView);
+        when(repositoryViewRepo.findById(1L)).thenReturn(Optional.of(mockRepositoryView));
 
         Object[] args = new Object[] { mockFedoraService, TEST_CONTEXT_ORG_URI };
         when(mockProceedingJoinPoint.getArgs()).thenReturn(args);
@@ -131,7 +132,7 @@ public class RepositoryViewContextControllerTest {
             delete(CONTROLLER_PATH, TEST_REPOSITORY_VIEW_TYPE, mockRepositoryView.getId())
                 .param("contextUri", TEST_CONTEXT_ORG_URI)
                 .content(objectMapper.writeValueAsString(mockRepositoryView))
-                .contentType(MediaType.APPLICATION_JSON_UTF8_VALUE
+                .contentType(MediaType.APPLICATION_JSON_VALUE
             )
         )
         .andExpect(status().isOk())
