@@ -1,6 +1,6 @@
 package edu.tamu.cap.controller.repositoryviewcontext;
 
-import static org.mockito.Matchers.any;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.when;
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
@@ -13,6 +13,8 @@ import static org.springframework.restdocs.request.RequestDocumentation.paramete
 import static org.springframework.restdocs.request.RequestDocumentation.pathParameters;
 import static org.springframework.restdocs.request.RequestDocumentation.requestParameters;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+import java.util.Optional;
 
 import org.aspectj.lang.ProceedingJoinPoint;
 import org.junit.jupiter.api.BeforeEach;
@@ -89,8 +91,8 @@ public class RepositoryViewContextQueryControllerTest {
         mockRepositoryView.setUsername("");
         mockRepositoryView.setPassword("");
 
-        when(repositoryViewRepo.getOne(1L)).thenReturn(mockRepositoryView);
-        when(repositoryViewRepo.findOne(1L)).thenReturn(mockRepositoryView);
+        when(repositoryViewRepo.getById(1L)).thenReturn(mockRepositoryView);
+        when(repositoryViewRepo.findById(1L)).thenReturn(Optional.of(mockRepositoryView));
 
         Object[] args = new Object[] { mockFedoraService, TEST_CONTEXT_ORG_URI };
         when(mockProceedingJoinPoint.getArgs()).thenReturn(args);
@@ -110,7 +112,7 @@ public class RepositoryViewContextQueryControllerTest {
             post(CONTROLLER_PATH, TEST_REPOSITORY_VIEW_TYPE, mockRepositoryView.getId())
                 .param("contextUri", TEST_CONTEXT_ORG_URI)
                 .content(objectMapper.writeValueAsString(TEST_QUERY))
-                .contentType(MediaType.APPLICATION_JSON_UTF8_VALUE)
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
         )
         .andExpect(status().isOk())
         .andDo(document("{method-name}/", preprocessRequest(prettyPrint()), preprocessResponse(prettyPrint()),
